server/requirements.txt
Flask==3.0.3
PyJWT==2.8.0
bcrypt==4.1.3
PyYAML==6.0.1
cryptography==42.0.8
pyopenssl==24.0.0
dtls==1.3.0
python-pytun==2.3.0
